requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
datasets==2.15.0
python-dotenv==1.0.0
flask==3.0.0
flask-session==0.5.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
bcrypt==4.0.1
pymongo==4.6.1
torch==2.1.2
transformers==4.36.2
faiss-cpu==1.7.4
spacy==3.7.2
openai==1.3.7
pyrebase4==4.7.1
python-jose==3.3.0
requests==2.31.0
